added correct parsing timestamps

+++ videoQueries/schemas/screenshots.py
@@ -10,7 +10,7 @@ class ScreenshotResponse(BaseModel):
     filename: str
     file_path: str
     timestamp_in_video: str
-    timestamp_in_seconds: int
+    timestamp_in_seconds: float
     created_at: Optional[datetime] = None  # Make it optional
     annotated_filename: str | None = None
     annotated_file_path: str | None = None
